requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
numpy==1.19.1
pandas==1.0.5
PyMySQL==0.10.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
xlrd==1.2.0
